+++ src/colnorms.h
@@ -1,5 +1,5 @@
 #ifndef COLNORMS_H
 #define COLNORMS_H
 #include<vector>
-std::vector<double> colnorms(std::vector<double>& d, std::vector<double>& lam, std::vector<double>& tau, std::vector<double>& org, std::vector<double>& v, double N = 17000);
+std::vector<double> colnorms(std::vector<double>& d, std::vector<double>& lam, std::vector<double>& tau, std::vector<int>& org, std::vector<double>& v, double N = 17000);
 #endif
